Split sector validation
